Split up PusherSwift.swift into smaller files
